Fix : delete unused file , favicon, html title, routing
