--- routes/commend.js
@@ -1,6 +1,0 @@
-const express = require('express');
-const controller = require('../controller/CCommend');
-const router = express.Router();
-
-
-module.exports = router;
