+++ views/404.ejs
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>404 Not Found</title>
+    <title>Card_Reomandation : 404 Page</title>
     <link rel="stylesheet" href="/css/404.css">
 </head>
 <body>
