--- views/commend.ejs
@@ -1,71 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <%- include('include/include') %>
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Card 3D view Parallax</title>
-    <link href="https://fonts.googleapis.com/css?family=Varela+Round" rel="stylesheet">
-    <link rel="stylesheet" href="/css/commend.css">
-</head>
-<body>
-    <div id="container">
-        <% cards.forEach(card => { %>
-            <div id="top">
-                <div class="perspective">
-                    <div class="card" id="card1">
-                        <!-- 카드 이미지 여기서 수정  -->
-                        <div class="thumb" data-url="/detail/<%= card.card_id %>" style="background-image: url(https://images.unsplash.com/photo-1478358161113-b0e11994a36b?ixlib=rb-0.3.5&ixid=eyJhcHBfaWQiOjEyMDd9&s=a11da19e81415e5a81029aade9df7774&auto=format&fit=crop&w=668&q=80)"></div>
-                        <h2><%= card.Card.card_name %></h2>
-                        <span><%= card[`card_${category}`] %></span>
-                    </div>
-                    <div class="comments">
-                        <% if (card.comments && card.comments.length > 0) { %>
-                            <!-- 상위 2개의 댓글만 가져오기 -->
-                            <% card.comments.slice(0, 2).forEach((comment, index) => { %>
-                            <div class="comment">
-                                <div class="comment-header">
-                                    <% if (index === 0) { %>
-                                        <span style="font-size: 40px;">🥇</span>
-                                    <% } else if (index === 1) { %>
-                                        <span style="font-size: 40px;">🥈</span>
-                                    <% } else { %>
-                                        <img src="https://via.placeholder.com/40" alt="Profile Image">
-                                    <% } %>
-                                    <div class="comment-author"><%= comment.User.nickname %></div>
-                                    <div class="comment-options"><%= comment.createdAt.toLocaleDateString() %></div>
-                                </div>
-                                <div class="comment-body" data-card-id="<%= card.card_id %>" data-comment-id="<%= comment.comment_id %>">
-                                    <%= comment.comment_contents %>
-                                </div>
-                                <div class="comment-footer">
-                                    <span class="likes">🔥</span> <!-- 좋아요 수를 나타내는 부분 -->
-                                    <span class="comments"><%= comment.dataValues.likeCount %></span> <!-- 댓글 수를 나타내는 부분 (현재 고정 값) -->
-                                    <span class="comment-action" data-likes="<%= comment.dataValues.likeCount %>">좋아요</span>
-                                </div>
-                            </div>
-                            <% }) %>
-                        <% } else { %>
-                            <div class="no-comments" data-url="/comment?card_id=<%= card.card_id %>">
-                                아직 입력된 댓글이 없습니다. 댓글 작성하기
-                            </div>
-                        <% } %>
-                    </div>
-                </div>
-            </div>
-        <% }) %>        
-    </div>
-
-    <!-- 모달 창 HTML 추가 -->
-    <div id="comment-modal" class="modal">
-        <div class="modal-content">
-            <span class="close">&times;</span>
-            <iframe id="comment-iframe" src="" frameborder="0"></iframe>
-        </div>
-    </div>
-
-    <script src="/js/commend.js"></script>
-    <script>
-    </script>
-</body>
-</html>
