--- views/comment.ejs
@@ -1,57 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <%- include('include/include') %>
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Comment Popup</title>
-    <link rel="stylesheet" href="/css/comment.css">
-</head>
-<body>
-    <h2>댓글 내용</h2>
-    <div class="comment-form">
-        <input type="text" id="comment-input" placeholder="댓글을 입력하세요">
-        <span class="char-counter" id="char-counter">0</span>
-        <button id="add-comment" data-card-id="<%= cardId %>">추가</button>
-    </div>
-
-    <div id="comments-container">
-        <!-- 댓글 목록이 여기에 추가됩니다 -->
-        <% if (comments.length === 0) { %>
-            <div class="no-comments">아직 입력된 댓글이 없습니다.</div>
-        <% } else { %>
-            <% comments.forEach((comment, index) => { %>
-                <% const isTopComment = index < 2; %> <!-- 상위 2개의 댓글을 식별 -->
-                <div class="comment <%= isTopComment ? 'top-comment' : '' %>" data-id="<%= comment.comment_id %>">
-                    <div class="comment-header">
-                        <div class="comment-author"><%= comment.User.nickname %></div>
-                        <div class="comment-options"><%= new Date(comment.createdAt).toLocaleDateString() %></div>
-                    </div>
-                    <div class="comment-body">
-                        <%= comment.comment_contents %>
-                    </div>
-                    <div class="comment-footer">
-                        <span class="likes"><%= comment.likeCount %></span> <!-- 좋아요 수를 나타내는 부분 -->
-                        <span class="comments">0</span> <!-- 댓글 수를 나타내는 부분 -->
-                        <span class="comment-action" data-comment-id="<%= comment.comment_id %>" data-card-id="<%= comment.card_id %>">좋아요</span> <!-- 좋아요 버튼에 card_id 추가 -->
-                        <% if (user && user.userId === comment.userid) { %>
-                            <button class="edit-comment">수정</button>
-                            <button class="delete-comment">삭제</button>
-                        <% } %>
-                    </div>
-                </div>
-            <% }) %>
-        <% } %>
-    </div>
-    
-    <!-- 페이지네이션 컨테이너 -->
-    <div class="pagination" id="pagination"></div>
-    <script>
-        let currentPage = <%= JSON.stringify(currentPage) %>;
-        let comments = <%- JSON.stringify(comments) %>;
-        let totalPages = <%= JSON.stringify(totalPages) %>;
-        const user = <%- JSON.stringify(user) %>;
-    </script>
-    <script src="/js/comment.js"></script>
-</body>
-</html>
