+++ views/cover.ejs
@@ -3,7 +3,8 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Animated Skills</title>
+    <title>Card_Reomandation : Cover</title>
+    <link rel="icon" href="/image/card.png">
     <link href="https://fonts.googleapis.com/css?family=Inconsolata:700" rel="stylesheet">
     <link rel="stylesheet" href="/css/cover.css">
 </head>
