+++ views/detail.ejs
@@ -4,7 +4,8 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Card Detail</title>
+    <title>Card_Reomandation : Detail</title>
+    <link rel="icon" href="/image/card.png">
     <link rel="stylesheet" href="/css/detail.css">
     <link rel="stylesheet" href="/css/comment.css">
 </head>
