+++ views/include/include.ejs
@@ -1,12 +1,8 @@
 <!-- views/include/include.ejs -->
-<head>
   <meta charset="UTF-8" />
   <meta http-equiv="X-UA-Compatible" content="IE=edge" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-
   <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
-</head>
-<body>
   <!-- 로그인 버튼과 관련된 섹션 제거 -->
   <script>
       function checkLoginStatus() {
@@ -25,4 +21,3 @@
           });
       });
   </script>
-</body>
