+++ views/index.ejs
@@ -2,8 +2,9 @@
 <html lang="en">
 <head>
     <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Expandable Items</title>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">    
+    <title>Card_Reomandation : index</title>  
+    <link rel="icon" href="/image/card.png">
     <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
     <link rel="stylesheet" href="/css/index.css">
 </head>
