+++ views/login.ejs
@@ -2,7 +2,8 @@
 <html lang="ko">
 <head>
   <meta charset="UTF-8">
-  <title>로그인</title>
+  <title>Card_Reomandation : Log-In</title>  
+  <link rel="icon" href="/image/card.png">
   <link rel="stylesheet" href="/css/login.css">
   <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
 </head>
