+++ views/profile.ejs
@@ -3,7 +3,8 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>마이 페이지</title>
+    <title>Card_Reomandation : My-Page</title>
+    <link rel="icon" href="/image/card.png">
     <link rel="stylesheet" href="/css/profile.css">
     <script src="https://cdn.jsdelivr.net/particles.js/2.0.0/particles.min.js"></script>
     <!-- Axios CDN 추가 -->
