+++ views/search.ejs
@@ -1,10 +1,11 @@
 <!DOCTYPE html>
 <html lang="en">
-    <%- include('include/include') %>
 <head>
+    <%- include('include/include') %>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Card Stack</title>
+    <title>Card_Reomandation : Search</title>
+    <link rel="icon" href="/image/card.png">
     <link rel="stylesheet" href="/css/search.css">
 </head>
 <body>
