+++ views/signup.ejs
@@ -2,7 +2,8 @@
 <html lang="ko">
 <head>
     <meta charset="UTF-8">
-    <title>회원가입</title>
+    <title>Card_Reomandation : Sign-Up</title>
+    <link rel="icon" href="/image/card.png">
     <link rel="stylesheet" href="/css/signup.css">
     <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
 </head>
